TrigonometriskPunkt -> man_made=survey_point
Map "TrigonometriskPunkt" to`man_made=survey_point` in addition to `natural=hill`, which is the current behaviour.

Was also debating with myself whether to add a `fixme` tag informing about the possibility of survey points not being on the top of a hill. But, if there are many survey points in a municipality, that might result in a lot of false positives to go through, so I decided against it for now. Any opinion on this?